Move Calculations to their own directory
