+++ packages/alfa-css/tsconfig.json
@@ -14,6 +14,7 @@
     "src/value.ts",
     "src/value/box.ts",
     "src/value/calculation.ts",
+    "src/value/calculation/calculation.ts",
     "src/value/color.ts",
     "src/value/color/current.ts",
     "src/value/color/hex.ts",
